Bugfixing

+++ Toolbox/scripts/util_clean.sh
@@ -1,23 +1,60 @@
 #!/bin/sh
+# Coded by Olli
+# This script will cleanup old stuff, which which isn't used anymore
+########################################################################################
 
-#Info
-DESCRIPTION="This script will cleanup old stuff, which isn't used anymore."
+# Info
+export DESCRIPTION="Starting cleanup of old stuff, which isn't used anymore"
+export MOUNTPOINT=1
 
 echo $DESCRIPTION
 
+# Include info script
 . /eso/hmi/engdefs/scripts/mqb/util_info.sh
-mount -uw /mnt/app
 
-echo "Deleting GreenMenus, which aren't used anymore."
-rm /mnt/app/eso/hmi/engdefs/mqb-adaptions.esd
-rm /mnt/app/eso/hmi/engdefs/mqb-rccAdaptions.esd
+# Include writeable system mount script
+. /eso/hmi/engdefs/scripts/mqb/util_mountsys.sh
 
-echo "Deleting scripts, which aren't used anymore."
-rm /mnt/app/eso/hmi/engdefs/scripts/mqb/install_online.sh
-rm /mnt/app/eso/hmi/engdefs/scripts/mqb/dump_online.sh
+echo "Deleting GreenMenus, which aren't used anymore..."
+if [ -f /mnt/app/eso/hmi/engdefs/mqb-adaptions.esd ]; then
+	rm /mnt/app/eso/hmi/engdefs/mqb-adaptions.esd
+	CLEANUP=yes
+fi
 
-mount -ur /mnt/app
+if [ -f /mnt/app/eso/hmi/engdefs/mqb-rccAdaptions.esd ]; then	
+	rm /mnt/app/eso/hmi/engdefs/mqb-rccAdaptions.esd
+	CLEANUP=yes
+fi
+sleep 1
 
-echo "Cleanup complete."
+echo "Deleting scripts, which aren't used anymore...."
+if [ -f /mnt/app/eso/hmi/engdefs/scripts/mqb/install_online.sh ]; then
+	rm /mnt/app/eso/hmi/engdefs/scripts/mqb/install_online.sh
+	CLEANUP=yes
+fi
+
+if [ -f /mnt/app/eso/hmi/engdefs/scripts/mqb/dump_online.sh ]; then
+	rm /mnt/app/eso/hmi/engdefs/scripts/mqb/dump_online.sh
+	CLEANUP=yes
+fi
+sleep 1
+
+if [ -f /mnt/app/eso/hmi/engdefs/scripts/mqb/set_VIM.sh ]; then
+	rm /mnt/app/eso/hmi/engdefs/scripts/mqb/set_VIM.sh
+	CLEANUP=yes
+fi
+sleep 1
+
+# Include back to read-only system mount script
+. /eso/hmi/engdefs/scripts/mqb/util_unmountsys.sh
+
+# Conclusion
+if [ "$CLEANUP" == "yes" ]; then
+	echo
+	echo "Cleanup complete"
+else
+	echo
+	echo "No old files found, nothing to cleanup"
+fi
 
 exit 0
